Storing data in the database

+++ app/src/main/java/com/example/dogsappjetpack/model/DogDao.java
@@ -10,7 +10,7 @@ import java.util.List;
 public interface DogDao {
 
     @Insert
-    List<Integer> insertAll(DogBreed...dogs);
+    List<Long> insertAll(DogBreed...dogs);
 
     @Query("SELECT * FROM dogbreed")
     List<DogBreed> getAllDogs();
